requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.199
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.165
    # via aws-cdk-lib
aws-cdk-aws-apigatewayv2-alpha==2.86.0.a0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
aws-cdk-aws-apigatewayv2-integrations-alpha==2.86.0a0
    # via -r requirements.in
aws-cdk-aws-lambda-python-alpha==2.86.0a0
    # via -r requirements.in
aws-cdk-aws-servicecatalogappregistry-alpha==2.86.0a0
    # via -r requirements.in
aws-cdk-lib==2.86.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-aws-servicecatalogappregistry-alpha
cattrs==23.1.2
    # via jsii
constructs==10.2.65
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-aws-servicecatalogappregistry-alpha
    #   aws-cdk-lib
exceptiongroup==1.1.1
    # via cattrs
importlib-resources==5.12.0
    # via jsii
jsii==1.84.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-aws-servicecatalogappregistry-alpha
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-aws-servicecatalogappregistry-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-aws-servicecatalogappregistry-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.7.0
    # via
    #   cattrs
    #   jsii
zipp==3.15.0
    # via importlib-resources
